+++ CHANGELOG.md
@@ -4,6 +4,10 @@ Please, check out guidelines: https://keepachangelog.com/en/1.0.0/
 
 ## Next version
 
+### Added
+
+- Install command https://github.com/tuist/tuist/pull/83 by @pepibumur.
+
 ## 0.1.0
 
 ### Added